Do not insert a fake elemet for length length into Lua hashes.
Original implementation would add an element with index "n"
and value equal to the hash length when converting from OCaml lists
to Lua hashes.

There was always a fallback to Luahash.length in the untable function,
and the fake element is guaranteed to interfere with iteration if
the rest of the list is something not compatible with numbers.